feat/quan 12 implement google ads on llc (#554)
* feat(llc):[QUAN-12]: add global gtag script

* feat(llc):[QUAN-12]: report conversion on form submission

* feat(llc/labs):[QUAN-12]: extend Window interface with gtag property

* feat(llc):[QUAN-12]: pass window.location object to gtag_report_conversion after form validation

* Remove URL from gtag_report_conversion call

We don't want to navigate to a new page as a result of the
conversion being reported. This was my (Brian's) mistake,
misunderstanding the function of the URL here.

* feat(llc):[QUAN-12]: after cr

* Re-remove argument to gtag_report_conversion()

We don't want the page to navigate anywhere after
the form is submitted. This fix was accidentally reverted
in the course of PR review.

* Add sourcing comment for gtag in _app.tsx

Co-authored-by: Brian Skinn <brian.skinn@gmail.com>