Apply review edits; trim whitespace from RAM fig.
Some spots for specific attention are called out with
bold-italic text bracketed in multiple exclamation points.

There was some extra whitespace at the top of the first figure
in the post (diagram of memory swapping), leading to a large gap
between the preceding text and the top of the figure. I cropped out
that extra whitespace.